requirements.txt
pandas
plotly_express
streamliT